moved parseIndicator to exchange added comments
