+++ gctscript/modules/ta/indicators/indicators_test.go
@@ -187,44 +187,3 @@ func TestToFloat64(t *testing.T) {
 		t.Fatalf("attempting to convert a string should fail but test passed")
 	}
 }
-
-func TestParseInterval(t *testing.T) {
-	v, err := ParseInterval("1h")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if v != time.Hour {
-		t.Fatalf("unexpected value return expected %v received %v", time.Hour, v)
-	}
-
-	v, err = ParseInterval("1d")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if v != time.Hour*24 {
-		t.Fatalf("unexpected value return expected %v received %v", time.Hour*24, v)
-	}
-
-	v, err = ParseInterval("3d")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if v != time.Hour*72 {
-		t.Fatalf("unexpected value return expected %v received %v", time.Hour*72, v)
-	}
-
-	v, err = ParseInterval("1w")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if v != time.Hour*168 {
-		t.Fatalf("unexpected value return expected %v received %v", time.Hour*72, v)
-	}
-
-	_, err = ParseInterval("6m")
-	if err != nil {
-		if !errors.Is(err, errInvalidInterval) {
-			t.Fatal(err)
-		}
-	}
-}
